+++ src/formats/mod.rs
@@ -1,4 +1,5 @@
 pub mod anvil;
+pub mod classic_schematic;
 pub mod litematic;
 pub mod manager;
 pub mod mcstructure;